SKERN-1616 Add to NCN DST build project
The node-identity feature has been identified as a package that
needs to be installed on Shasta NCN nodes. This commit adds the
rpm produced by the DST pipeline into the Shasta CD Repo (aka "Blob")
for installation.

The node-identity is expected to be installed on the NCN whether
Shasta Premium or Shasta Standard, so Jenkinsfile has both in the
PRODUCT field.

UPDATE: Corrected target_os setting in Jenkinsfile. It appears to be
an no-op but want to set per DST guidance.

Test Description:
Verified the node-identity rpm is built and published to the SLES 15
NCN repo in the SKERN CAR repo. The copy of the rpm into the blob
is not controlled by SKERN but prior packages designated for the NCN
reside in the same folder and are in the Blob.

Target Release: None